+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.6.0
+
+- Added option to use SPOT instances for execution nodes.
+- Implement optional creation of IVS buckets
+
 ## v0.5.0
 
 - Implemented deployment of an OpenSearch Domain for usage in IVS